requirements.txt
streamlit
Pillow
mysql-connector-python
pandas
python-dotenv
googletrans==4.0.0-rc1
google-generativeai